# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Mindful AI Sandbox API
  version: 1.0.0
  description: |
    Mindful AI Sandbox is a student learning environment for building and experimenting
    with cloud-native applications on AWS.

    **Key Features:**
    - User authentication and role-based access control (User, Admin)
    - User profile management
    - Admin user management

    **Authentication:**
    Most endpoints require JWT authentication via the `Authorization` header.
    Use the `/auth/login` endpoint to obtain an access token.

  contact:
    name: Mindful AI Sandbox Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.example.com/dev
    description: Development server
  - url: https://api.example.com/staging
    description: Staging server
  - url: https://api.example.com/prod
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      description: Create a new user account. Default role is User.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              user:
                summary: Register as User
                value:
                  email: user@example.com
                  password: SecurePass123!
                  username: student42
                  role: User
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate user
      description: Login with email and password to receive access and refresh tokens.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: user@example.com
              password: SecurePass123!
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Use refresh token to obtain a new access token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            example:
              refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Logout user
      description: Invalidate user session (client-side token removal).
      security: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/confirm-signup:
    post:
      tags: [Authentication]
      summary: Confirm email address
      description: Verify email address using confirmation code sent via email.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmSignUpRequest'
            example:
              email: user@example.com
              confirmationCode: '123456'
      responses:
        '200':
          description: Email confirmed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Request password reset
      description: Send password reset code to user's email.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
            example:
              email: user@example.com
      responses:
        '200':
          description: Password reset code sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/confirm-forgot-password:
    post:
      tags: [Authentication]
      summary: Reset password
      description: Reset password using confirmation code and new password.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmForgotPasswordRequest'
            example:
              email: user@example.com
              confirmationCode: '123456'
              newPassword: NewSecurePass123!
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

  # ============================================================================
  # PROFILE ENDPOINTS
  # ============================================================================
  /me:
    get:
      tags: [Profile]
      summary: Get current user profile
      description: Retrieve authenticated user's private profile (includes email).
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateProfile'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags: [Profile]
      summary: Update current user profile
      description: Update profile information (username, bio, avatar).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
            example:
              username: student42
              bio: Learning cloud-native development
              avatarUrl: https://example.com/avatar.jpg
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateProfileResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login
  schemas:
...
    UserRole:
      type: string
      enum: [User, Admin]
      description: User role in the system
...
    RegisterRequest:
      type: object
      required: [email, password, username]
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          format: password
          minLength: 8
          description: User password (min 8 chars, must include uppercase, lowercase, and number)
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[a-zA-Z0-9_-]+$'
          description: Unique username (alphanumeric, underscore, hyphen only)
        role:
          $ref: '#/components/schemas/UserRole'
          default: User
    RegisterResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponseWrapper'
        - type: object
          properties:
            data:
              type: object
              properties:
                userId:
                  type: string
                  example: "38510330-9041-70d5-010c-a30b00bd812d"
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                username:
                  type: string
                  example: "student42"
                role:
                  $ref: '#/components/schemas/UserRole'
                message:
                  type: string
                  example: "User registered successfully. Please check your email for confirmation code."
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponseWrapper'
        - type: object
          properties:
            data:
              type: object
              required: [user, tokens]
              properties:
                user:
                  type: object
                  properties:
                    userId:
                      type: string
                      example: "38510330-9041-70d5-010c-a30b00bd812d"
                    email:
                      type: string
                      format: email
                      example: "user@example.com"
                    username:
                      type: string
                      example: "test_user"
                    role:
                      $ref: '#/components/schemas/UserRole'
                    createdAt:
                      type: string
                      format: date-time
                      description: User account creation timestamp
                      example: "2025-10-28T06:35:13.807Z"
                tokens:
                  type: object
                  required: [accessToken, refreshToken, expiresIn, tokenType]
                  properties:
                    accessToken:
                      type: string
                      description: JWT access token (expires in 15 minutes)
                      example: "eyJraWQiOi..."
                    refreshToken:
                      type: string
                      description: JWT refresh token (expires in 7 days)
                      example: "eyJjdHkiOi..."
                    idToken:
                      type: string
                      description: Cognito ID token
                      example: "eyJraWQiOi..."
                    expiresIn:
                      type: integer
                      description: Access token expiration time in seconds
                      example: 900
                    tokenType:
                      type: string
                      description: Token type (always Bearer)
                      example: "Bearer"
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          description: Valid refresh token
    RefreshTokenResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponseWrapper'
        - type: object
          properties:
            data:
              type: object
              properties:
                accessToken:
                  type: string
                  description: New JWT access token
                  example: "eyJraWQiOi..."
                expiresIn:
                  type: integer
                  description: Access token expiration time in seconds
                  example: 900
                tokenType:
                  type: string
                  description: Token type (always Bearer)
                  example: "Bearer"
    ConfirmSignUpRequest:
      type: object
      required: [email, confirmationCode]
      properties:
        email:
          type: string
          format: email
        confirmationCode:
          type: string
          minLength: 6
          maxLength: 6
          description: 6-digit confirmation code sent via email
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ConfirmForgotPasswordRequest:
      type: object
      required: [email, confirmationCode, newPassword]
      properties:
        email:
          type: string
          format: email
        confirmationCode:
          type: string
          minLength: 6
          maxLength: 6
        newPassword:
          type: string
          format: password
          minLength: 8

    # =========================================================================
    # PROFILE SCHEMAS
    # =========================================================================
    UpdateProfileRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[a-zA-Z0-9_-]+$'
        bio:
          type: string
          maxLength: 500
        avatarUrl:
          type: string
          format: uri
    PrivateProfile:
      allOf:
        - $ref: '#/components/schemas/SuccessResponseWrapper'
        - type: object
          properties:
            data:
              type: object
              properties:
                userId:
                  type: string
                  example: "38510330-9041-70d5-010c-a30b00bd812d"
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                username:
                  type: string
                  example: "test_user"
                role:
                  $ref: '#/components/schemas/UserRole'
                bio:
                  type: string
                  nullable: true
                  example: "Learning cloud-native development"
                avatarUrl:
                  type: string
                  nullable: true
                  example: "https://example.com/avatar.jpg"
                createdAt:
                  type: string
                  format: date-time
                  example: "2025-10-28T06:35:13.807Z"
                updatedAt:
                  type: string
                  format: date-time
                  example: "2025-11-01T21:03:23.065Z"
...
    UpdateProfileResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponseWrapper'
        - type: object
          properties:
            data:
              type: object
              properties:
                message:
                  type: string
                  example: "Profile updated successfully"
                profile:
                  type: object
                  description: Updated profile data (structure same as PrivateProfile data field)

    # =========================================================================
    # ADMIN USER MANAGEMENT SCHEMAS
    # =========================================================================
...
    ResponseMeta:
      type: object
      required: [timestamp, requestId]
      properties:
        timestamp:
          type: string
          format: date-time
          description: Server timestamp when response was generated
          example: "2025-11-01T21:03:23.065Z"
        requestId:
          type: string
          format: uuid
          description: Unique request ID for debugging and tracing
          example: "ca1fec38-e588-4be7-acbc-921575189fbe"
    SuccessResponseWrapper:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          description: Indicates successful response
          example: true
        data:
          description: Response payload (structure varies by endpoint, can be object or array)
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ErrorResponseWrapper:
      type: object
      required: [success, error, meta]
      properties:
        success:
          type: boolean
          description: Indicates failed response
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Error code for client-side error handling
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
              example: "Invalid input data"
            details:
              type: object
              nullable: true
              description: Additional error details (field-level errors, etc.)
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    MessageResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponseWrapper'
        - type: object
          properties:
            data:
              type: object
              properties:
                message:
                  type: string
                  example: "Operation completed successfully"
...
  responses:
    ValidationError:
      description: Validation error (invalid request data)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseWrapper'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Validation failed
              details:
                validationErrors:
                  - field: email
                    message: Invalid email format
            meta:
              timestamp: "2025-11-01T21:03:23.065Z"
              requestId: "ca1fec38-e588-4be7-acbc-921575189fbe"
    UnauthorizedError:
      description: Unauthorized (missing or invalid authentication)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseWrapper'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
            meta:
              timestamp: "2025-11-01T21:03:23.065Z"
              requestId: "ca1fec38-e588-4be7-acbc-921575189fbe"
...
    ConflictError:
      description: Conflict (resource already exists or state conflict)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseWrapper'
          example:
            success: false
            error:
              code: CONFLICT
              message: Resource already exists
            meta:
              timestamp: "2025-11-01T21:03:23.065Z"
              requestId: "ca1fec38-e588-4be7-acbc-921575189fbe"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseWrapper'
          example:
            success: false
            error:
              code: INTERNAL_ERROR
              message: An unexpected error occurred
            meta:
              timestamp: "2025-11-01T21:03:23.065Z"
              requestId: "ca1fec38-e588-4be7-acbc-921575189fbe"